create stream folder and move stream-related files into it
