+++ components/script/dom/byteteereadintorequest.rs
@@ -19,7 +19,7 @@ use crate::dom::bindings::root::Dom;
 use crate::dom::byteteeunderlyingsource::ByteTeeUnderlyingSource;
 use crate::dom::globalscope::GlobalScope;
 use crate::dom::promise::Promise;
-use crate::dom::readablestream::ReadableStream;
+use crate::dom::stream::readablestream::ReadableStream;
 use crate::microtask::Microtask;
 use crate::script_runtime::CanGc;
