+++ components/script/dom/byteteeunderlyingsource.rs
@@ -23,7 +23,7 @@ use crate::dom::bindings::root::Dom;
 use crate::dom::byteteereadrequest::ByteTeeReadRequest;
 use crate::dom::globalscope::GlobalScope;
 use crate::dom::promise::Promise;
-use crate::dom::readablestreamdefaultreader::ReadRequest;
+use crate::dom::stream::readablestreamdefaultreader::ReadRequest;
 use crate::script_runtime::{CanGc, JSContext as SafeJSContext};
 
 #[derive(JSTraceable, MallocSizeOf)]
