+++ components/script/dom/mod.rs
@@ -243,7 +243,6 @@ pub(crate) mod closeevent;
 pub(crate) mod commandevent;
 pub(crate) mod comment;
 pub(crate) mod compositionevent;
-pub(crate) mod compressionstream;
 pub(crate) mod console;
 pub(crate) mod cookiestore;
 pub(crate) mod countqueuingstrategy;
@@ -265,7 +264,6 @@ pub(crate) mod debuggergetpossiblebreakpointsevent;
 pub(crate) mod debuggerglobalscope;
 pub(crate) mod debuggerpauseevent;
 pub(crate) mod debuggersetbreakpointevent;
-pub(crate) mod decompressionstream;
 pub(crate) mod defaultteereadrequest;
 pub(crate) mod defaultteeunderlyingsource;
 pub(crate) mod dissimilaroriginlocation;
@@ -375,13 +373,6 @@ pub(crate) mod radionodelist;
 pub(crate) mod range;
 pub(crate) mod raredata;
 #[expect(dead_code)]
-pub(crate) mod readablebytestreamcontroller;
-pub(crate) mod readablestream;
-pub(crate) mod readablestreambyobreader;
-pub(crate) mod readablestreambyobrequest;
-pub(crate) mod readablestreamdefaultcontroller;
-pub(crate) mod readablestreamdefaultreader;
-pub(crate) mod readablestreamgenericreader;
 pub(crate) mod reportingendpoint;
 pub(crate) mod reportingobserver;
 pub(crate) mod request;
@@ -401,6 +392,7 @@ pub(crate) mod shadowroot;
 pub(crate) mod staticrange;
 pub(crate) mod storage;
 pub(crate) mod storageevent;
+pub(crate) mod stream;
 pub(crate) mod submitevent;
 pub(crate) mod subtlecrypto;
 pub(crate) mod svg;
@@ -413,9 +405,7 @@ pub(crate) mod text;
 pub(crate) mod textcontrol;
 pub(crate) mod textdecoder;
 pub(crate) mod textdecodercommon;
-pub(crate) mod textdecoderstream;
 pub(crate) mod textencoder;
-pub(crate) mod textencoderstream;
 pub(crate) mod texttrack;
 pub(crate) mod texttrackcue;
 pub(crate) mod texttrackcuelist;
@@ -466,8 +456,6 @@ pub(crate) use self::webgpu::*;
 pub(crate) mod gpucanvascontext;
 pub(crate) mod webrtc;
 pub(crate) use self::webrtc::*;
-pub(crate) mod transformstream;
-pub(crate) mod transformstreamdefaultcontroller;
 pub(crate) mod wheelevent;
 #[expect(dead_code)]
 pub(crate) mod window;
@@ -476,9 +464,6 @@ pub(crate) mod workers;
 pub(crate) use self::workers::*;
 pub(crate) mod worklet;
 pub(crate) mod workletglobalscope;
-pub(crate) mod writablestream;
-pub(crate) mod writablestreamdefaultcontroller;
-pub(crate) mod writablestreamdefaultwriter;
 pub(crate) mod xmldocument;
 pub(crate) mod xmlhttprequest;
 pub(crate) mod xmlhttprequesteventtarget;
