+++ components/script/dom/response.rs
@@ -31,7 +31,7 @@ use crate::dom::bindings::str::{ByteString, USVString, serialize_jsval_to_json_u
 use crate::dom::globalscope::GlobalScope;
 use crate::dom::headers::{Guard, Headers, is_obs_text, is_vchar};
 use crate::dom::promise::Promise;
-use crate::dom::readablestream::ReadableStream;
+use crate::dom::stream::readablestream::ReadableStream;
 use crate::dom::underlyingsourcecontainer::UnderlyingSourceType;
 use crate::script_runtime::{CanGc, JSContext, StreamConsumer};
