+++ components/script/dom/stream/compressionstream.rs
@@ -25,7 +25,7 @@ use crate::dom::bindings::conversions::{SafeFromJSValConvertible, SafeToJSValCon
 use crate::dom::bindings::error::{Error, Fallible};
 use crate::dom::bindings::reflector::{Reflector, reflect_dom_object_with_proto};
 use crate::dom::bindings::root::{Dom, DomRoot};
-use crate::dom::transformstreamdefaultcontroller::TransformerType;
+use crate::dom::stream::transformstreamdefaultcontroller::TransformerType;
 use crate::dom::types::{
     GlobalScope, ReadableStream, TransformStream, TransformStreamDefaultController, WritableStream,
 };
