+++ components/script/dom/stream/mod.rs
@@ -1,0 +1,16 @@
+pub(crate) mod compressionstream;
+pub(crate) mod decompressionstream;
+pub(crate) mod readablebytestreamcontroller;
+pub(crate) mod readablestream;
+pub(crate) mod readablestreambyobreader;
+pub(crate) mod readablestreambyobrequest;
+pub(crate) mod readablestreamdefaultcontroller;
+pub(crate) mod readablestreamdefaultreader;
+pub(crate) mod readablestreamgenericreader;
+pub(crate) mod textdecoderstream;
+pub(crate) mod textencoderstream;
+pub(crate) mod transformstream;
+pub(crate) mod transformstreamdefaultcontroller;
+pub(crate) mod writablestream;
+pub(crate) mod writablestreamdefaultcontroller;
+pub(crate) mod writablestreamdefaultwriter;
