+++ components/script/dom/stream/readablebytestreamcontroller.rs
@@ -33,8 +33,8 @@ use crate::dom::bindings::trace::RootedTraceableBox;
 use crate::dom::globalscope::GlobalScope;
 use crate::dom::promise::Promise;
 use crate::dom::promisenativehandler::{Callback, PromiseNativeHandler};
-use crate::dom::readablestream::ReadableStream;
-use crate::dom::readablestreambyobrequest::ReadableStreamBYOBRequest;
+use crate::dom::stream::readablestream::ReadableStream;
+use crate::dom::stream::readablestreambyobrequest::ReadableStreamBYOBRequest;
 use crate::realms::{InRealm, enter_realm};
 use crate::script_runtime::{CanGc, JSContext as SafeJSContext};
