+++ components/script/dom/stream/readablestreamdefaultcontroller.rs
@@ -29,8 +29,8 @@ use crate::dom::bindings::trace::RootedTraceableBox;
 use crate::dom::globalscope::GlobalScope;
 use crate::dom::promise::Promise;
 use crate::dom::promisenativehandler::{Callback, PromiseNativeHandler};
-use crate::dom::readablestream::ReadableStream;
-use crate::dom::readablestreamdefaultreader::ReadRequest;
+use crate::dom::stream::readablestream::ReadableStream;
+use crate::dom::stream::readablestreamdefaultreader::ReadRequest;
 use crate::dom::underlyingsourcecontainer::{UnderlyingSourceContainer, UnderlyingSourceType};
 use crate::realms::{InRealm, enter_realm};
 use crate::script_runtime::{CanGc, JSContext as SafeJSContext};
