+++ components/script/dom/stream/readablestreamdefaultreader.rs
@@ -31,7 +31,7 @@ use crate::dom::globalscope::GlobalScope;
 use crate::dom::promise::Promise;
 use crate::dom::promisenativehandler::{Callback, PromiseNativeHandler};
 use crate::dom::readablestream::{ReadableStream, bytes_from_chunk_jsval};
-use crate::dom::readablestreamgenericreader::ReadableStreamGenericReader;
+use crate::dom::stream::readablestreamgenericreader::ReadableStreamGenericReader;
 use crate::realms::{InRealm, enter_realm};
 use crate::script_runtime::{CanGc, JSContext as SafeJSContext};
