+++ components/script/dom/stream/readablestreamgenericreader.rs
@@ -14,8 +14,8 @@ use crate::dom::bindings::reflector::DomGlobal;
 use crate::dom::bindings::root::{DomRoot, MutNullableDom};
 use crate::dom::globalscope::GlobalScope;
 use crate::dom::promise::Promise;
-use crate::dom::readablestreambyobreader::ReadableStreamBYOBReader;
-use crate::dom::readablestreamdefaultreader::ReadableStreamDefaultReader;
+use crate::dom::stream::readablestreambyobreader::ReadableStreamBYOBReader;
+use crate::dom::stream::readablestreamdefaultreader::ReadableStreamDefaultReader;
 use crate::script_runtime::{CanGc, JSContext as SafeJSContext};
 
 /// <https://streams.spec.whatwg.org/#readablestreamgenericreader>
