+++ components/script/dom/stream/writablestreamdefaultwriter.rs
@@ -15,7 +15,7 @@ use crate::dom::bindings::reflector::{DomGlobal, Reflector, reflect_dom_object_w
 use crate::dom::bindings::root::{DomRoot, MutNullableDom};
 use crate::dom::globalscope::GlobalScope;
 use crate::dom::promise::Promise;
-use crate::dom::writablestream::WritableStream;
+use crate::dom::stream::writablestream::WritableStream;
 use crate::realms::InRealm;
 use crate::script_runtime::{CanGc, JSContext as SafeJSContext};
