+++ components/script/dom/xmlhttprequest.rs
@@ -66,9 +66,9 @@ use crate::dom::headers::is_forbidden_request_header;
 use crate::dom::node::Node;
 use crate::dom::performance::performanceresourcetiming::InitiatorType;
 use crate::dom::progressevent::ProgressEvent;
-use crate::dom::readablestream::ReadableStream;
 use crate::dom::servoparser::ServoParser;
 use crate::dom::servoparser::html::HtmlSerialize;
+use crate::dom::stream::readablestream::ReadableStream;
 use crate::dom::window::Window;
 use crate::dom::workerglobalscope::WorkerGlobalScope;
 use crate::dom::xmlhttprequesteventtarget::XMLHttpRequestEventTarget;
